refactor(commands): split chat.rs into per-responsibility submodules
Splits the 3,239-line src-tauri/src/commands/chat.rs monolith into a
chat/ directory with focused submodules: send, lifecycle, checkpoint,
naming, interaction, attachments, session, plus shared helpers in
mod.rs. main.rs now references commands via their owning submodule
(e.g. chat::send::send_chat_message) so generate_handler! can resolve
the macro-generated __cmd__ items. No behavioral changes.

Closes #498
