+++ src-tauri/src/commands/chat/send.rs
@@ -1,185 +1,44 @@
 use std::sync::Arc;
 
-use serde::{Deserialize, Serialize};
 use tauri::{AppHandle, Emitter, Manager, State};
 
 use claudette::agent::{
     self, AgentEvent, AgentSettings, ControlRequestInner, FileAttachment, InnerStreamEvent,
     PersistentSession, StartContentBlock, StreamEvent,
 };
+use claudette::base64_decode;
 use claudette::db::Database;
 use claudette::env::WorkspaceEnv;
-use claudette::git;
 use claudette::mcp_supervisor::McpSupervisor;
-use claudette::model::{
-    Attachment, ChatMessage, ChatRole, ChatSession, CompletedTurnData, ConversationCheckpoint,
-    TurnToolActivity,
-};
-use claudette::snapshot;
-use claudette::{base64_decode, base64_encode};
-
-use crate::agent_mcp_sink::ChatBridgeSink;
-use crate::state::{AgentSessionState, AppState, PendingPermission};
-use claudette::agent_mcp::bridge::{BridgeHandle, McpBridgeSession};
-
-/// Build a fresh bridge for a workspace and return an `mcp_config` JSON with
-/// the synthetic `claudette` MCP server entry merged in. The Claude CLI will
-/// spawn `claudette-tauri --agent-mcp` as a stdio child of itself and pass it
-/// the per-session socket address + token via env vars; the grandchild then
-/// connects back to the parent over the local socket.
-async fn start_bridge_and_inject_mcp(
-    app: &AppHandle,
-    db_path: &std::path::Path,
-    workspace_id: &str,
-    chat_session_id: &str,
-    base_mcp_config: Option<String>,
-) -> Result<(Arc<McpBridgeSession>, Option<String>), String> {
-    let sink = Arc::new(ChatBridgeSink {
-        app: app.clone(),
-        db_path: db_path.to_path_buf(),
-        workspace_id: workspace_id.to_string(),
-        chat_session_id: chat_session_id.to_string(),
-    });
-    let bridge = Arc::new(McpBridgeSession::start(sink).await?);
-    let merged = inject_claudette_mcp_entry(base_mcp_config, bridge.handle())?;
-    Ok((bridge, merged))
-}
-
-fn inject_claudette_mcp_entry(
-    base: Option<String>,
-    handle: &BridgeHandle,
-) -> Result<Option<String>, String> {
-    let exe = std::env::current_exe()
-        .map_err(|e| format!("current_exe: {e}"))?
-        .to_string_lossy()
-        .to_string();
-
-    let entry = serde_json::json!({
-        "type": "stdio",
-        "command": exe,
-        "args": ["--agent-mcp"],
-        "env": {
-            claudette::agent_mcp::server::ENV_SOCKET_ADDR: handle.socket_addr,
-            claudette::agent_mcp::server::ENV_TOKEN: handle.token,
-        }
-    });
-
-    let mut wrapper: serde_json::Value = match base.as_deref() {
-        Some(s) => serde_json::from_str(s).map_err(|e| format!("parse mcp_config: {e}"))?,
-        None => serde_json::json!({"mcpServers": {}}),
-    };
-    let servers = wrapper
-        .get_mut("mcpServers")
-        .and_then(|v| v.as_object_mut())
-        .ok_or_else(|| "mcp_config missing `mcpServers` object".to_string())?;
-    servers.insert(claudette::agent_mcp::server::SERVER_NAME.to_string(), entry);
-    Ok(Some(wrapper.to_string()))
-}
-
-/// How long to wait between emitting `agent-permission-prompt` and firing the
-/// attention system notification. This is the window in which the webview
-/// picks up the event, runs the Zustand setter, and paints the question/plan
-/// card. 300ms is a compromise: long enough to cover typical React render
-/// plus a macOS window-show animation (when the user had the window hidden),
-/// short enough that the notification still feels tied to the trigger.
-const ATTENTION_NOTIFY_DELAY_MS: u64 = 300;
-
-async fn fire_completion_notification(
-    db_path: &std::path::Path,
-    cesp_playback: &std::sync::Mutex<claudette::cesp::SoundPlaybackState>,
-    event: crate::tray::NotificationEvent,
-    ws_id: &str,
-) {
-    let Ok(db) = Database::open(db_path) else {
-        return;
-    };
-    let resolved = crate::tray::resolve_notification(&db, cesp_playback, event);
-    if resolved.sound != "None" {
-        crate::commands::settings::play_notification_sound(resolved.sound, Some(resolved.volume));
-    }
-    if let Ok(Some(cmd)) = db.get_app_setting("notification_command")
-        && !cmd.is_empty()
-        && let Some(fresh_ws) = db
-            .list_workspaces()
-            .ok()
-            .and_then(|wss| wss.into_iter().find(|w| w.id == ws_id))
-    {
-        let repo = db.get_repository(&fresh_ws.repository_id).ok().flatten();
-        let repo_path = repo.as_ref().map(|r| r.path.as_str()).unwrap_or_default();
-        let default_branch = match repo.as_ref().and_then(|r| r.base_branch.as_deref()) {
-            Some(b) => b.to_string(),
-            None => git::default_branch(
-                repo_path,
-                repo.as_ref().and_then(|r| r.default_remote.as_deref()),
-            )
-            .await
-            .unwrap_or_else(|_| "main".into()),
-        };
-        let fresh_env = WorkspaceEnv::from_workspace(&fresh_ws, repo_path, default_branch);
-        if let Some(mut command) =
-            crate::commands::settings::build_notification_command(&cmd, &fresh_env)
-            && let Ok(child) = command.spawn()
-        {
-            crate::commands::settings::spawn_and_reap(child);
-        }
-    }
-}
-
-/// Frontend-facing input for a file attachment (base64-encoded).
-#[derive(Clone, Deserialize)]
-pub struct AttachmentInput {
-    pub filename: String,
-    pub media_type: String,
-    pub data_base64: String,
-    pub text_content: Option<String>,
-}
-
-/// Frontend-facing response for a stored attachment (base64-encoded data).
-#[derive(Clone, Serialize)]
-pub struct AttachmentResponse {
-    pub id: String,
-    pub message_id: String,
-    pub filename: String,
-    pub media_type: String,
-    pub data_base64: String,
-    pub text_content: Option<String>,
-    pub width: Option<i32>,
-    pub height: Option<i32>,
-    pub size_bytes: i64,
-    /// `"user"` for composer-supplied attachments, `"agent"` for ones the
-    /// agent delivered via `mcp__claudette__send_to_user`. The frontend
-    /// uses this to re-route agent attachments under the assistant message
-    /// instead of the user message they were FK-anchored to. Without this
-    /// on reload, persisted agent attachments display as "from you".
-    pub origin: claudette::model::AttachmentOrigin,
-    pub tool_use_id: Option<String>,
-}
-
-#[derive(Clone, Serialize)]
-struct AgentStreamPayload {
-    workspace_id: String,
-    chat_session_id: String,
-    event: AgentEvent,
-}
-
+use claudette::model::{ChatMessage, ChatRole, ConversationCheckpoint};
 use claudette::permissions::{is_bypass_tools, tools_for_level};
+use claudette::snapshot;
+
+use crate::state::{AgentSessionState, AppState, PendingPermission};
+
+use super::interaction::{deny_drained_permissions, drain_pending_permissions};
+use super::naming::{try_auto_rename, try_generate_session_name};
+use super::{
+    ATTENTION_NOTIFY_DELAY_MS, AgentStreamPayload, AttachmentInput, fire_completion_notification,
+    now_iso, start_bridge_and_inject_mcp,
+};
 
 /// Spawn-time flags of the currently running persistent session, plus the
 /// backend-observed `exited_plan` latch (set when the agent emits
 /// `ExitPlanMode` during this session).
-struct SessionFlags<'a> {
-    plan_mode: bool,
-    allowed_tools: &'a [String],
-    exited_plan: bool,
-    disable_1m_context: bool,
+pub(crate) struct SessionFlags<'a> {
+    pub plan_mode: bool,
+    pub allowed_tools: &'a [String],
+    pub exited_plan: bool,
+    pub disable_1m_context: bool,
 }
 
 /// Flags the next turn is asking for. Compared against [`SessionFlags`] to
 /// decide whether the process must be torn down and respawned.
-struct RequestedFlags<'a> {
-    plan_mode: bool,
-    allowed_tools: &'a [String],
-    disable_1m_context: bool,
+pub(crate) struct RequestedFlags<'a> {
+    pub plan_mode: bool,
+    pub allowed_tools: &'a [String],
+    pub disable_1m_context: bool,
 }
 
 /// Detect whether the persistent session's spawn-time flags have drifted
@@ -193,7 +52,7 @@ struct RequestedFlags<'a> {
 /// `plan_mode`, the plan phase is over regardless of whether the frontend
 /// remembered to send `plan_mode=false` — force a teardown so the CLI
 /// respawns without `--permission-mode plan`.
-fn persistent_session_flags_drifted(
+pub(crate) fn persistent_session_flags_drifted(
     session: SessionFlags<'_>,
     requested: RequestedFlags<'_>,
 ) -> bool {
@@ -227,7 +86,7 @@ fn persistent_session_flags_drifted(
 /// Auto-allow in bypass mode does not bypass an MCP server's own
 /// authorization — servers refuse at their layer via a normal tool_result,
 /// not a control_request.
-fn build_permission_response(
+pub(crate) fn build_permission_response(
     session_allowed_tools: &[String],
     session_plan_mode: bool,
     session_exited_plan: bool,
@@ -354,7 +213,7 @@ pub async fn send_chat_message(
         )
     }
 
-    let mut att_models: Vec<Attachment> = Vec::new();
+    let mut att_models: Vec<claudette::model::Attachment> = Vec::new();
     let mut cli_atts: Vec<FileAttachment> = Vec::new();
 
     if let Some(ref inputs) = attachments {
@@ -405,7 +264,7 @@ pub async fn send_chat_message(
                 None
             };
             let size_bytes = data.len() as i64;
-            att_models.push(Attachment {
+            att_models.push(claudette::model::Attachment {
                 id: uuid::Uuid::new_v4().to_string(),
                 message_id: user_msg.id.clone(),
                 filename: input.filename.clone(),
@@ -713,7 +572,7 @@ pub async fn send_chat_message(
     let repo_path = repo.as_ref().map(|r| r.path.as_str()).unwrap_or("");
     let default_branch = match repo.as_ref().and_then(|r| r.base_branch.as_deref()) {
         Some(b) => b.to_string(),
-        None => git::default_branch(
+        None => claudette::git::default_branch(
             repo_path,
             repo.as_ref().and_then(|r| r.default_remote.as_deref()),
         )
@@ -1768,1133 +1627,11 @@ pub async fn send_chat_message(
     Ok(())
 }
 
-/// What `take_stop_snapshot` hands back: permissions to deny, pid to kill,
-/// and the session id to mark ended in the DB.
-type StopSnapshot = (
-    Option<(Arc<PersistentSession>, Vec<PendingPermission>)>,
-    Option<u32>,
-    Option<String>,
-);
-
-/// Mutations performed on an agent session when the user clicks Stop.
-///
-/// Stop interrupts the in-flight turn — it takes `active_pid` so the caller
-/// can kill the process and drains pending permission requests so the caller
-/// can deny them. It deliberately does NOT clear `session_id`, `turn_count`,
-/// or `persistent_session`: those are owned by `reset_agent_session` and
-/// `clear_conversation`. Preserving them is what lets the next
-/// `send_chat_message` resume via `--resume` instead of spawning a fresh
-/// conversation. The now-dead `persistent_session` handle is fine — on the
-/// next turn `send_turn` detects the broken pipe and respawns with
-/// `--resume <session_id>`.
-fn take_stop_snapshot(session: &mut AgentSessionState) -> StopSnapshot {
-    let drained = drain_pending_permissions(session);
-    let ended_sid = Some(session.session_id.clone());
-    let pid = session.active_pid.take();
-    session.needs_attention = false;
-    session.attention_kind = None;
-    (drained, pid, ended_sid)
-}
-
-#[tauri::command]
-pub async fn stop_agent(
-    session_id: String,
-    app: AppHandle,
-    state: State<'_, AppState>,
-) -> Result<(), String> {
-    let db = Database::open(&state.db_path).map_err(|e| e.to_string())?;
-    let chat_session_id = session_id;
-    let chat_session = db
-        .get_chat_session(&chat_session_id)
-        .map_err(|e| e.to_string())?
-        .ok_or("Chat session not found")?;
-    let workspace_id = chat_session.workspace_id.clone();
-
-    // Drain pending permissions and snapshot the cleanup state synchronously
-    // under the lock; deny sends, the kill, and the DB session-end happen
-    // after we release it.
-    let (to_deny_stop, pid_to_kill, ended_sid) = {
-        let mut agents = state.agents.write().await;
-        match agents.get_mut(&chat_session_id) {
-            Some(session) => take_stop_snapshot(session),
-            None => (None, None, None),
-        }
-    };
-
-    if let Some((ref ps, drained)) = to_deny_stop {
-        deny_drained_permissions(drained, ps, "Session stopped by user.").await;
-    }
-    if let Some(pid) = pid_to_kill {
-        agent::stop_agent(pid).await?;
-    }
-
-    // Stop aborts an in-flight turn but deliberately preserves the persisted
-    // Claude resume UUID and turn count so the next turn can continue the
-    // conversation via `--resume`. Only the agent_sessions audit row is closed
-    // out as a failure.
-    if let Some(sid) = ended_sid.as_deref().filter(|s| !s.is_empty()) {
-        let _ = db.end_agent_session(sid, false);
-    }
-
-    crate::tray::rebuild_tray(&app);
-
-    // Log stop message on this chat session.
-    let msg = ChatMessage {
-        id: uuid::Uuid::new_v4().to_string(),
-        workspace_id,
-        chat_session_id,
-        role: ChatRole::System,
-        content: "Agent stopped".to_string(),
-        cost_usd: None,
-        duration_ms: None,
-        created_at: now_iso(),
-        thinking: None,
-        input_tokens: None,
-        output_tokens: None,
-        cache_read_tokens: None,
-        cache_creation_tokens: None,
-    };
-    db.insert_chat_message(&msg).map_err(|e| e.to_string())?;
-
-    Ok(())
-}
-
-#[tauri::command]
-pub async fn reset_agent_session(
-    session_id: String,
-    app: AppHandle,
-    state: State<'_, AppState>,
-) -> Result<(), String> {
-    let db = Database::open(&state.db_path).map_err(|e| e.to_string())?;
-    let chat_session_id = session_id;
-
-    // Drain pending permissions under the lock, remove the session, capture
-    // its session_id and any active PID; the deny sends, process kill,
-    // and DB session-end happen after release.
-    let (to_deny_reset, ended_sid, pid_to_kill) = {
-        let mut agents = state.agents.write().await;
-        let drained = agents
-            .get_mut(&chat_session_id)
-            .and_then(drain_pending_permissions);
-        let removed = agents.remove(&chat_session_id);
-        let ended_sid = removed.as_ref().map(|s| s.session_id.clone());
-        let pid_to_kill = removed.and_then(|s| s.active_pid);
-        (drained, ended_sid, pid_to_kill)
-    };
-
-    if let Some((ref ps, drained)) = to_deny_reset {
-        deny_drained_permissions(drained, ps, "Session reset.").await;
-    }
-    if let Some(pid) = pid_to_kill {
-        let _ = agent::stop_agent(pid).await;
-    }
-
-    // Clear persisted claude state so the next turn starts fresh. Reset
-    // discards in-flight state, so record as a failure.
-    db.clear_chat_session_state(&chat_session_id)
-        .map_err(|e| e.to_string())?;
-    if let Some(sid) = ended_sid.as_deref().filter(|s| !s.is_empty()) {
-        let _ = db.end_agent_session(sid, false);
-    }
-
-    crate::tray::rebuild_tray(&app);
-    Ok(())
-}
-
-#[tauri::command]
-pub async fn list_checkpoints(
-    session_id: String,
-    state: State<'_, AppState>,
-) -> Result<Vec<ConversationCheckpoint>, String> {
-    let db = Database::open(&state.db_path).map_err(|e| e.to_string())?;
-    db.list_checkpoints_for_session(&session_id)
-        .map_err(|e| e.to_string())
-}
-
-#[tauri::command]
-pub async fn rollback_to_checkpoint(
-    session_id: String,
-    checkpoint_id: String,
-    restore_files: bool,
-    state: State<'_, AppState>,
-) -> Result<Vec<ChatMessage>, String> {
-    let db = Database::open(&state.db_path).map_err(|e| e.to_string())?;
-    let chat_session_id = session_id;
-
-    // Load the target checkpoint and verify ownership.
-    let checkpoint = db
-        .get_checkpoint(&checkpoint_id)
-        .map_err(|e| e.to_string())?
-        .ok_or("Checkpoint not found")?;
-    // Scope the rollback to the given session — prevents a rollback in tab A
-    // from pruning messages in tab B. For pre-v20 checkpoints (empty
-    // chat_session_id) we accept the caller's chat_session_id as
-    // authoritative, since the backfill would have assigned all such
-    // messages to a single session.
-    if !checkpoint.chat_session_id.is_empty() && checkpoint.chat_session_id != chat_session_id {
-        return Err("Checkpoint does not belong to this session".into());
-    }
-
-    // Look up the workspace for the file-restore branch below.
-    let chat_session = db
-        .get_chat_session(&chat_session_id)
-        .map_err(|e| e.to_string())?
-        .ok_or("Chat session not found")?;
-    let workspace_id = chat_session.workspace_id.clone();
-
-    // Guard: reject if agent is running.
-    {
-        let agents = state.agents.read().await;
-        if let Some(session) = agents.get(&chat_session_id)
-            && session.active_pid.is_some()
-        {
-            return Err("Cannot rollback while the agent is running".into());
-        }
-    }
-
-    // Attempt file restore BEFORE any destructive DB writes so that a
-    // failure does not leave the DB truncated while the frontend still shows
-    // the full conversation.
-    if restore_files {
-        let workspaces = db.list_workspaces().map_err(|e| e.to_string())?;
-        let ws = workspaces
-            .iter()
-            .find(|w| w.id == workspace_id)
-            .ok_or("Workspace not found")?;
-        let wt = ws
-            .worktree_path
-            .as_ref()
-            .ok_or("Workspace has no worktree")?;
-
-        if checkpoint.has_file_state {
-            // New path: restore from SQLite snapshot.
-            snapshot::restore_snapshot(&state.db_path, &checkpoint_id, wt)
-                .await
-                .map_err(|e| e.to_string())?;
-        } else if let Some(ref commit_hash) = checkpoint.commit_hash {
-            // Legacy path: restore from git commit.
-            git::restore_to_commit(wt, commit_hash)
-                .await
-                .map_err(|e| e.to_string())?;
-        }
-    }
-
-    // Now perform the destructive DB writes — safe because the risky git
-    // operation (if requested) has already succeeded above. Scoped to this
-    // session so sibling tabs are untouched.
-    db.delete_session_messages_after(&chat_session_id, &checkpoint.message_id)
-        .map_err(|e| e.to_string())?;
-    db.delete_session_checkpoints_after(&chat_session_id, checkpoint.turn_index)
-        .map_err(|e| e.to_string())?;
-
-    // Reset the per-session agent state so the next turn starts fresh.
-    // Rollback discards the session's prior work — record as a failure.
-    let ended_sid = {
-        let mut agents = state.agents.write().await;
-        agents.remove(&chat_session_id).map(|s| s.session_id)
-    };
-    if let Some(sid) = ended_sid.as_deref() {
-        let _ = db.end_agent_session(sid, false);
-    }
-    db.clear_chat_session_state(&chat_session_id)
-        .map_err(|e| e.to_string())?;
-
-    // Return the truncated message list for this session.
-    db.list_chat_messages_for_session(&chat_session_id)
-        .map_err(|e| e.to_string())
-}
-
-/// Clear the entire conversation for a workspace, optionally restoring files
-/// to the merge-base (initial state before any agent work).
-#[tauri::command]
-pub async fn clear_conversation(
-    session_id: String,
-    restore_files: bool,
-    state: State<'_, AppState>,
-) -> Result<Vec<ChatMessage>, String> {
-    let db = Database::open(&state.db_path).map_err(|e| e.to_string())?;
-    let chat_session_id = session_id;
-    let chat_session = db
-        .get_chat_session(&chat_session_id)
-        .map_err(|e| e.to_string())?
-        .ok_or("Chat session not found")?;
-    let workspace_id = chat_session.workspace_id.clone();
-
-    // Guard: reject if agent is running.
-    {
-        let agents = state.agents.read().await;
-        if let Some(session) = agents.get(&chat_session_id)
-            && session.active_pid.is_some()
-        {
-            return Err("Cannot clear conversation while the agent is running".into());
-        }
-    }
-
-    // Optionally restore files to the merge-base before clearing.
-    if restore_files {
-        let workspaces = db.list_workspaces().map_err(|e| e.to_string())?;
-        let ws = workspaces
-            .iter()
-            .find(|w| w.id == workspace_id)
-            .ok_or("Workspace not found")?;
-        let wt = ws
-            .worktree_path
-            .as_ref()
-            .ok_or("Workspace has no worktree")?;
-        let repos = db.list_repositories().map_err(|e| e.to_string())?;
-        let repo = repos
-            .iter()
-            .find(|r| r.id == ws.repository_id)
-            .ok_or("Repository not found")?;
-        let base = match repo.base_branch.as_deref() {
-            Some(b) => b.to_string(),
-            None => git::default_branch(&repo.path, repo.default_remote.as_deref())
-                .await
-                .map_err(|e| e.to_string())?,
-        };
-        let merge_base = claudette::diff::merge_base(wt, &ws.branch_name, &base)
-            .await
-            .map_err(|e| e.to_string())?;
-        git::restore_to_commit(wt, &merge_base)
-            .await
-            .map_err(|e| e.to_string())?;
-    }
-
-    // Delete all messages and checkpoints for this session.
-    db.delete_chat_messages_for_session(&chat_session_id)
-        .map_err(|e| e.to_string())?;
-    // Checkpoints cascade via FK, but delete explicitly for clarity.
-    db.delete_session_checkpoints_after(&chat_session_id, -1)
-        .map_err(|e| e.to_string())?;
-
-    // Reset agent session. Clearing the conversation discards the session's
-    // prior work, so it did not run to completion — record as a failure.
-    let ended_sid = {
-        let mut agents = state.agents.write().await;
-        agents.remove(&chat_session_id).map(|s| s.session_id)
-    };
-    if let Some(sid) = ended_sid.as_deref() {
-        let _ = db.end_agent_session(sid, false);
-    }
-    db.clear_chat_session_state(&chat_session_id)
-        .map_err(|e| e.to_string())?;
-
-    // Return empty list.
-    db.list_chat_messages_for_session(&chat_session_id)
-        .map_err(|e| e.to_string())
-}
-
-#[tauri::command]
-pub async fn save_turn_tool_activities(
-    checkpoint_id: String,
-    message_count: i32,
-    activities: Vec<TurnToolActivity>,
-    state: State<'_, AppState>,
-) -> Result<(), String> {
-    let db = Database::open(&state.db_path).map_err(|e| e.to_string())?;
-    db.save_turn_tool_activities(&checkpoint_id, message_count, &activities)
-        .map_err(|e| e.to_string())?;
-    Ok(())
-}
-
-#[tauri::command]
-pub async fn load_completed_turns(
-    session_id: String,
-    state: State<'_, AppState>,
-) -> Result<Vec<CompletedTurnData>, String> {
-    let db = Database::open(&state.db_path).map_err(|e| e.to_string())?;
-    db.list_completed_turns_for_session(&session_id)
-        .map_err(|e| e.to_string())
-}
-
-/// Background task: generate a descriptive branch name via Haiku and rename
-/// the workspace's branch + DB record. All failures are non-fatal.
-#[allow(clippy::too_many_arguments)]
-async fn try_auto_rename(
-    ws_id: &str,
-    worktree_path: &str,
-    old_name: &str,
-    old_branch: &str,
-    prompt: &str,
-    branch_rename_preferences: Option<&str>,
-    db_path: &std::path::Path,
-    app: &AppHandle,
-    ws_env: &WorkspaceEnv,
-) {
-    // Ask Haiku for a branch name slug.
-    let slug = match agent::generate_branch_name(
-        prompt,
-        worktree_path,
-        branch_rename_preferences,
-        Some(ws_env),
-    )
-    .await
-    {
-        Ok(s) => s,
-        Err(_) => return,
-    };
-
-    // Resolve the configured branch prefix.
-    let prefix = {
-        let db = match Database::open(db_path) {
-            Ok(db) => db,
-            Err(_) => return,
-        };
-        let (mode, custom) = super::workspace::read_branch_prefix_settings(&db);
-        // Drop db before the async call (Database is not Sync).
-        drop(db);
-        super::workspace::resolve_branch_prefix(&mode, &custom).await
-    };
-
-    // Try the slug, then slug-2, slug-3 on name collision.
-    let candidates = [slug.clone(), format!("{slug}-2"), format!("{slug}-3")];
-    for candidate in &candidates {
-        let new_branch = format!("{prefix}{candidate}");
-
-        let db = match Database::open(db_path) {
-            Ok(db) => db,
-            Err(_) => return,
-        };
-
-        match db.rename_workspace(ws_id, candidate, &new_branch) {
-            Ok(()) => {
-                // DB updated — now rename the git branch.
-                if let Err(e) = git::rename_branch(worktree_path, old_branch, &new_branch).await {
-                    let _ = db.rename_workspace(ws_id, old_name, old_branch);
-
-                    // If the target branch already exists, fall back to the next
-                    // candidate just like we do for DB unique constraint collisions.
-                    if e.to_string().contains("already exists") {
-                        continue;
-                    }
-                    return;
-                }
-
-                // Success — notify the frontend.
-                let payload = serde_json::json!({
-                    "workspace_id": ws_id,
-                    "name": candidate,
-                    "branch_name": new_branch,
-                });
-                let _ = app.emit("workspace-renamed", &payload);
-                return;
-            }
-            Err(e) => {
-                if e.to_string().contains("UNIQUE constraint failed") {
-                    continue;
-                }
-                return;
-            }
-        }
-    }
-}
-
-/// Background task: ask Haiku for a short session name and persist it. All
-/// failures are non-fatal — if Haiku is unavailable or the user has already
-/// renamed the session, the `New chat` default stays in place.
-async fn try_generate_session_name(
-    session_id: &str,
-    worktree_path: &str,
-    prompt: &str,
-    db_path: &std::path::Path,
-    app: &AppHandle,
-    ws_env: &WorkspaceEnv,
-) {
-    let name = match agent::generate_session_name(prompt, worktree_path, Some(ws_env)).await {
-        Ok(n) => n,
-        Err(_) => return,
-    };
-
-    let db = match Database::open(db_path) {
-        Ok(db) => db,
-        Err(_) => return,
-    };
-
-    // The helper only writes when `name_edited == 0`, so a concurrent user
-    // rename between spawn and write is handled correctly — we become a no-op.
-    if let Ok(true) = db.set_session_name_from_haiku(session_id, &name) {
-        let payload = serde_json::json!({
-            "session_id": session_id,
-            "name": name,
-        });
-        let _ = app.emit("session-renamed", &payload);
-    }
-}
-
-#[tauri::command]
-pub async fn clear_attention(
-    session_id: String,
-    app: AppHandle,
-    state: State<'_, AppState>,
-) -> Result<(), String> {
-    let chat_session_id = session_id;
-
-    let mut agents = state.agents.write().await;
-    if let Some(session) = agents.get_mut(&chat_session_id)
-        && session.needs_attention
-    {
-        session.needs_attention = false;
-        drop(agents);
-        crate::tray::rebuild_tray(&app);
-    }
-    Ok(())
-}
-
-/// Resolve a pending AskUserQuestion `can_use_tool` request with the user's
-/// answers. `answers` is keyed by question text (matching the CLI's
-/// `mapToolResultToToolResultBlockParam` expectation) and layered onto the
-/// original tool input as `updatedInput`. The CLI then runs the tool's
-/// `call(updatedInput)` which produces the real tool_result.
-#[tauri::command]
-pub async fn submit_agent_answer(
-    session_id: String,
-    tool_use_id: String,
-    answers: std::collections::HashMap<String, String>,
-    annotations: Option<serde_json::Value>,
-    state: State<'_, AppState>,
-) -> Result<(), String> {
-    // Validate everything BEFORE removing the pending entry: if the session
-    // has been torn down or the entry maps to the wrong tool, the entry must
-    // stay so the user (or the correct submit_* command) can still see it.
-    let (pending, ps) = {
-        let mut agents = state.agents.write().await;
-        let session = agents.get_mut(&session_id).ok_or("Session not found")?;
-        // 1. Persistent session must be alive — otherwise nobody is reading
-        //    stdin and the response would be discarded.
-        let ps = session
-            .persistent_session
-            .clone()
-            .ok_or("Agent session is not active")?;
-        // 2. Tool kind must match — peek by reference.
-        match session.pending_permissions.get(&tool_use_id) {
-            None => {
-                let pending_ids: Vec<String> =
-                    session.pending_permissions.keys().cloned().collect();
-                return Err(format!(
-                    "No pending permission request for tool_use_id {tool_use_id} (pending: {pending_ids:?})"
-                ));
-            }
-            Some(p) if p.tool_name != "AskUserQuestion" => {
-                return Err(format!(
-                    "Pending tool for {tool_use_id} is {}, not AskUserQuestion",
-                    p.tool_name
-                ));
-            }
-            _ => {}
-        }
-        // 3. All checks passed — now it is safe to remove.
-        let pending = session
-            .pending_permissions
-            .remove(&tool_use_id)
-            .expect("checked above");
-        session.needs_attention = false;
-        session.attention_kind = None;
-        session.attention_notification_sent = false;
-        (pending, ps)
-    };
-
-    // Layer answers (and annotations, if any) onto the original input.
-    let mut updated_input = pending.original_input.clone();
-    if !updated_input.is_object() {
-        updated_input = serde_json::Value::Object(serde_json::Map::new());
-    }
-    if let Some(obj) = updated_input.as_object_mut() {
-        let answers_value =
-            serde_json::to_value(&answers).map_err(|e| format!("Failed to encode answers: {e}"))?;
-        obj.insert("answers".to_string(), answers_value);
-        if let Some(ann) = annotations {
-            obj.insert("annotations".to_string(), ann);
-        }
-    }
-
-    let response = serde_json::json!({
-        "behavior": "allow",
-        "updatedInput": updated_input,
-    });
-    ps.send_control_response(&pending.request_id, response)
-        .await
-}
-
-/// Resolve a pending ExitPlanMode `can_use_tool` request.
-/// `approved=true` → allow with the model's original input (the CLI's
-/// `call()` will save the plan and emit the real tool_result).
-/// `approved=false` → deny with the given reason (or a sensible default).
-#[tauri::command]
-pub async fn submit_plan_approval(
-    session_id: String,
-    tool_use_id: String,
-    approved: bool,
-    reason: Option<String>,
-    state: State<'_, AppState>,
-) -> Result<(), String> {
-    // Same validate-before-remove pattern as submit_agent_answer — see that
-    // function for the rationale.
-    let (pending, ps) = {
-        let mut agents = state.agents.write().await;
-        let session = agents.get_mut(&session_id).ok_or("Session not found")?;
-        let ps = session
-            .persistent_session
-            .clone()
-            .ok_or("Agent session is not active")?;
-        match session.pending_permissions.get(&tool_use_id) {
-            None => {
-                let pending_ids: Vec<String> =
-                    session.pending_permissions.keys().cloned().collect();
-                return Err(format!(
-                    "No pending permission request for tool_use_id {tool_use_id} (pending: {pending_ids:?})"
-                ));
-            }
-            Some(p) if p.tool_name != "ExitPlanMode" => {
-                return Err(format!(
-                    "Pending tool for {tool_use_id} is {}, not ExitPlanMode",
-                    p.tool_name
-                ));
-            }
-            _ => {}
-        }
-        let pending = session
-            .pending_permissions
-            .remove(&tool_use_id)
-            .expect("checked above");
-        session.needs_attention = false;
-        session.attention_kind = None;
-        session.attention_notification_sent = false;
-        (pending, ps)
-    };
-
-    let response = if approved {
-        serde_json::json!({
-            "behavior": "allow",
-            "updatedInput": pending.original_input,
-        })
-    } else {
-        let feedback = reason.unwrap_or_else(|| "Plan denied. Please revise the approach.".into());
-        let message = format!(
-            "{feedback}\n\nRevise the plan to address this feedback, then call ExitPlanMode again to present the updated plan for approval. Do not begin implementation until the user approves the revised plan."
-        );
-        serde_json::json!({
-            "behavior": "deny",
-            "message": message,
-        })
-    };
-    ps.send_control_response(&pending.request_id, response)
-        .await
-}
-
-/// Synchronously drain any pending permission requests from `session` and
-/// snapshot the [`PersistentSession`] needed to deny them. Designed to be
-/// called while holding the agents write lock — does no async work itself.
-///
-/// Returns `None` when there is nothing to do (no pending entries) or when
-/// there is no live `PersistentSession` to receive the denies (entries are
-/// dropped in that case, since nobody could read the response anyway).
-fn drain_pending_permissions(
-    session: &mut AgentSessionState,
-) -> Option<(Arc<PersistentSession>, Vec<PendingPermission>)> {
-    if session.pending_permissions.is_empty() {
-        return None;
-    }
-    let Some(ps) = session.persistent_session.clone() else {
-        session.pending_permissions.clear();
-        return None;
-    };
-    let drained: Vec<PendingPermission> = session
-        .pending_permissions
-        .drain()
-        .map(|(_, p)| p)
-        .collect();
-    Some((ps, drained))
-}
-
-/// Send a deny `control_response` for each drained permission. Caller must
-/// have already released the agents lock — this performs async I/O against
-/// the CLI's stdin and would otherwise serialize all other agent-state ops.
-async fn deny_drained_permissions(
-    drained: Vec<PendingPermission>,
-    ps: &PersistentSession,
-    reason: &str,
-) {
-    for pending in drained {
-        let deny = serde_json::json!({
-            "behavior": "deny",
-            "message": reason,
-        });
-        if let Err(e) = ps.send_control_response(&pending.request_id, deny).await {
-            eprintln!(
-                "[chat] Failed to deny pending {} on cleanup: {e}",
-                pending.tool_name
-            );
-        }
-    }
-}
-
-/// Load attachment metadata for a single session's chat history.
-///
-/// Images (< ~5 MB base64) include inline data for immediate rendering.
-/// Documents (PDFs, potentially 20+ MB) omit the body — use
-/// [`load_attachment_data`] to fetch on demand.
-#[tauri::command]
-pub async fn load_attachments_for_session(
-    session_id: String,
-    state: State<'_, AppState>,
-) -> Result<Vec<AttachmentResponse>, String> {
-    let db = Database::open(&state.db_path).map_err(|e| e.to_string())?;
-    let messages = db
-        .list_chat_messages_for_session(&session_id)
-        .map_err(|e| e.to_string())?;
-    let message_ids: Vec<String> = messages.iter().map(|m| m.id.clone()).collect();
-    let att_map = db
-        .list_attachments_for_messages(&message_ids)
-        .map_err(|e| e.to_string())?;
-
-    let mut result = Vec::new();
-    for (_, atts) in att_map {
-        for a in atts {
-            // Inline base64 data for images and text-shaped attachments —
-            // PDFs are too large to push through IPC eagerly and would stall
-            // the renderer.
-            let is_text = matches!(
-                a.media_type.as_str(),
-                "text/plain" | "text/csv" | "text/markdown" | "application/json"
-            );
-            let data_base64 = if a.media_type.starts_with("image/") || is_text {
-                base64_encode(&a.data)
-            } else {
-                String::new()
-            };
-            let text_content = if is_text {
-                std::str::from_utf8(&a.data).ok().map(str::to_owned)
-            } else {
-                None
-            };
-            result.push(AttachmentResponse {
-                id: a.id,
-                message_id: a.message_id,
-                filename: a.filename,
-                media_type: a.media_type,
-                data_base64,
-                text_content,
-                width: a.width,
-                height: a.height,
-                size_bytes: a.size_bytes,
-                origin: a.origin,
-                tool_use_id: a.tool_use_id,
-            });
-        }
-    }
-    Ok(result)
-}
-
-/// Fetch the full base64-encoded body of a single attachment by ID.
-/// Used for on-demand loading of large attachments (PDFs).
-#[tauri::command]
-pub async fn load_attachment_data(
-    attachment_id: String,
-    state: State<'_, AppState>,
-) -> Result<String, String> {
-    let db = Database::open(&state.db_path).map_err(|e| e.to_string())?;
-    let att = db
-        .get_attachment(&attachment_id)
-        .map_err(|e| e.to_string())?
-        .ok_or("Attachment not found")?;
-    Ok(base64_encode(&att.data))
-}
-
-/// Read a file from disk and return it as base64 with metadata.
-/// Used by the frontend file picker — avoids needing the `plugin-fs` dependency.
-///
-/// Known types (images, PDFs) use their specific MIME types and size limits.
-/// Unknown extensions are tested for UTF-8 validity — valid text files are
-/// returned as `text/plain` with the content in `text_content`.
-#[tauri::command]
-pub async fn read_file_as_base64(path: String) -> Result<AttachmentResponse, String> {
-    use std::path::Path;
-
-    const MAX_IMAGE_SIZE: usize = 3_932_160; // 3.75 MB
-    const MAX_PDF_SIZE: usize = 20 * 1024 * 1024; // 20 MB
-    const MAX_TEXT_SIZE: usize = 1024 * 1024; // 1 MB
-    const MAX_CSV_SIZE: usize = 2 * 1024 * 1024; // 2 MB
-    const MAX_MARKDOWN_SIZE: usize = 1024 * 1024; // 1 MB
-    const MAX_JSON_SIZE: usize = 1024 * 1024; // 1 MB
-
-    let file_path = Path::new(&path);
-    let filename = file_path
-        .file_name()
-        .and_then(|n| n.to_str())
-        .unwrap_or("attachment")
-        .to_string();
-
-    let ext = file_path
-        .extension()
-        .and_then(|e| e.to_str())
-        .unwrap_or("")
-        .to_lowercase();
-
-    enum Kind {
-        Image,
-        Pdf,
-        Csv,
-        Markdown,
-        Json,
-    }
-
-    // Map extensions to a Kind. Anything else falls through to UTF-8
-    // sniffing below, which classifies the file as plain text or rejects
-    // it as binary.
-    let known_kind = match ext.as_str() {
-        "png" => Some((Kind::Image, "image/png")),
-        "jpg" | "jpeg" => Some((Kind::Image, "image/jpeg")),
-        "gif" => Some((Kind::Image, "image/gif")),
-        "webp" => Some((Kind::Image, "image/webp")),
-        "svg" => Some((Kind::Image, "image/svg+xml")),
-        "pdf" => Some((Kind::Pdf, "application/pdf")),
-        "csv" => Some((Kind::Csv, "text/csv")),
-        "md" | "markdown" => Some((Kind::Markdown, "text/markdown")),
-        "json" => Some((Kind::Json, "application/json")),
-        _ => None,
-    };
-
-    // Check file size via metadata before reading to avoid loading huge
-    // files into memory only to reject them.
-    let metadata = tokio::fs::metadata(&path)
-        .await
-        .map_err(|e| format!("Failed to read file: {e}"))?;
-    let file_len = metadata.len() as usize;
-    if file_len == 0 {
-        return Err("File is empty".to_string());
-    }
-    let max_for_read = match &known_kind {
-        Some((Kind::Pdf, _)) => MAX_PDF_SIZE,
-        Some((Kind::Image, _)) => MAX_IMAGE_SIZE,
-        Some((Kind::Csv, _)) => MAX_CSV_SIZE,
-        Some((Kind::Markdown, _)) => MAX_MARKDOWN_SIZE,
-        Some((Kind::Json, _)) => MAX_JSON_SIZE,
-        None => MAX_TEXT_SIZE,
-    };
-    if file_len > max_for_read {
-        return Err(format!(
-            "File too large: {} (max {})",
-            humanize_size(file_len),
-            humanize_size(max_for_read)
-        ));
-    }
-
-    let data = tokio::fs::read(&path)
-        .await
-        .map_err(|e| format!("Failed to read file: {e}"))?;
-
-    let size_bytes = data.len() as i64;
-
-    match known_kind {
-        Some((Kind::Image, media_type)) => {
-            let data_base64 = base64_encode(&data);
-            Ok(AttachmentResponse {
-                id: String::new(),
-                message_id: String::new(),
-                filename,
-                media_type: media_type.to_string(),
-                data_base64,
-                text_content: None,
-                width: None,
-                height: None,
-                size_bytes,
-                origin: claudette::model::AttachmentOrigin::User,
-                tool_use_id: None,
-            })
-        }
-        Some((Kind::Pdf, media_type)) => {
-            if !data.starts_with(b"%PDF-") {
-                return Err("Invalid PDF file: missing %PDF- header".to_string());
-            }
-            let data_base64 = base64_encode(&data);
-            Ok(AttachmentResponse {
-                id: String::new(),
-                message_id: String::new(),
-                filename,
-                media_type: media_type.to_string(),
-                data_base64,
-                text_content: None,
-                width: None,
-                height: None,
-                size_bytes,
-                origin: claudette::model::AttachmentOrigin::User,
-                tool_use_id: None,
-            })
-        }
-        Some((Kind::Csv | Kind::Markdown | Kind::Json, media_type)) => {
-            // Text-shaped — must be valid UTF-8 *and* contain no NUL bytes.
-            // UTF-8 itself permits U+0000, so a binary file with a misleading
-            // `.csv` / `.json` / `.md` extension can pass `from_utf8` while
-            // clearly being binary. The unknown-extension branch below
-            // already rejects on `\0`; mirror that here so the same content
-            // doesn't get a free pass purely because the suffix is on the
-            // known list.
-            let check_len = data.len().min(8192);
-            if data[..check_len].contains(&0) {
-                return Err(format!("File looks binary (contains NUL bytes): .{ext}"));
-            }
-            let text = String::from_utf8(data.clone())
-                .map_err(|_| format!("File is not valid UTF-8: .{ext}"))?;
-            let data_base64 = base64_encode(&data);
-            Ok(AttachmentResponse {
-                id: String::new(),
-                message_id: String::new(),
-                filename,
-                media_type: media_type.to_string(),
-                data_base64,
-                text_content: Some(text),
-                width: None,
-                height: None,
-                size_bytes,
-                origin: claudette::model::AttachmentOrigin::User,
-                tool_use_id: None,
-            })
-        }
-        None => {
-            // Unknown extension — attempt to read as text.
-            let check_len = data.len().min(8192);
-            if data[..check_len].contains(&0) {
-                return Err(format!("Unsupported binary file type: .{ext}"));
-            }
-            match String::from_utf8(data.clone()) {
-                Ok(text) => {
-                    let data_base64 = base64_encode(&data);
-                    Ok(AttachmentResponse {
-                        id: String::new(),
-                        message_id: String::new(),
-                        filename,
-                        media_type: "text/plain".to_string(),
-                        data_base64,
-                        text_content: Some(text),
-                        width: None,
-                        height: None,
-                        size_bytes,
-                        origin: claudette::model::AttachmentOrigin::User,
-                        tool_use_id: None,
-                    })
-                }
-                Err(_) => Err(format!("Unsupported binary file type: .{ext}")),
-            }
-        }
-    }
-}
-
-fn humanize_size(bytes: usize) -> String {
-    if bytes >= 1_048_576 {
-        format!("{:.1} MB", bytes as f64 / 1_048_576.0)
-    } else {
-        format!("{} KB", bytes / 1024)
-    }
-}
-
-fn now_iso() -> String {
-    let dur = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default();
-    format!("{}", dur.as_secs())
-}
-
-/// Overlay live runtime fields (agent_status, needs_attention, attention_kind)
-/// from `AppState.agents` onto a `ChatSession` loaded from the DB. The DB
-/// defaults are `Idle`/`false`/`None`; this replaces them with whatever the
-/// running agent map reports.
-fn hydrate_session(
-    mut session: ChatSession,
-    agents: &std::collections::HashMap<String, AgentSessionState>,
-) -> ChatSession {
-    if let Some(agent) = agents.get(&session.id) {
-        session.agent_status = if agent.active_pid.is_some() {
-            claudette::model::AgentStatus::Running
-        } else {
-            claudette::model::AgentStatus::Idle
-        };
-        session.needs_attention = agent.needs_attention;
-        session.attention_kind = agent.attention_kind.map(|k| match k {
-            crate::state::AttentionKind::Ask => claudette::model::AttentionKind::Ask,
-            crate::state::AttentionKind::Plan => claudette::model::AttentionKind::Plan,
-        });
-    }
-    session
-}
-
-#[tauri::command]
-pub async fn list_chat_sessions(
-    workspace_id: String,
-    include_archived: Option<bool>,
-    state: State<'_, AppState>,
-) -> Result<Vec<ChatSession>, String> {
-    let db = Database::open(&state.db_path).map_err(|e| e.to_string())?;
-    let sessions = db
-        .list_chat_sessions_for_workspace(&workspace_id, include_archived.unwrap_or(false))
-        .map_err(|e| e.to_string())?;
-    let agents = state.agents.read().await;
-    Ok(sessions
-        .into_iter()
-        .map(|s| hydrate_session(s, &agents))
-        .collect())
-}
-
-#[tauri::command]
-pub async fn get_chat_session(
-    session_id: String,
-    state: State<'_, AppState>,
-) -> Result<ChatSession, String> {
-    let db = Database::open(&state.db_path).map_err(|e| e.to_string())?;
-    let session = db
-        .get_chat_session(&session_id)
-        .map_err(|e| e.to_string())?
-        .ok_or("Session not found")?;
-    let agents = state.agents.read().await;
-    Ok(hydrate_session(session, &agents))
-}
-
-#[tauri::command]
-pub async fn create_chat_session(
-    workspace_id: String,
-    state: State<'_, AppState>,
-) -> Result<ChatSession, String> {
-    let db = Database::open(&state.db_path).map_err(|e| e.to_string())?;
-    let session = db
-        .create_chat_session(&workspace_id)
-        .map_err(|e| e.to_string())?;
-    let agents = state.agents.read().await;
-    Ok(hydrate_session(session, &agents))
-}
-
-#[tauri::command]
-pub async fn rename_chat_session(
-    session_id: String,
-    name: String,
-    state: State<'_, AppState>,
-) -> Result<(), String> {
-    let capped = claudette::model::validate_session_name(&name).map_err(String::from)?;
-    let db = Database::open(&state.db_path).map_err(|e| e.to_string())?;
-    db.rename_chat_session(&session_id, &capped)
-        .map_err(|e| e.to_string())
-}
-
-/// Archive a chat session (soft-delete). Stops its running agent first, then
-/// marks the row archived. If this was the workspace's last active session,
-/// a fresh `New chat` session is created so every workspace always has ≥1
-/// active session. Returns the newly created session in that case, `None`
-/// otherwise — the frontend uses the return value to select the new tab.
-#[tauri::command]
-pub async fn archive_chat_session(
-    app: AppHandle,
-    session_id: String,
-    state: State<'_, AppState>,
-) -> Result<Option<ChatSession>, String> {
-    let db = Database::open(&state.db_path).map_err(|e| e.to_string())?;
-    let session = db
-        .get_chat_session(&session_id)
-        .map_err(|e| e.to_string())?
-        .ok_or("Session not found")?;
-    let workspace_id = session.workspace_id.clone();
-
-    // Stop and remove the live agent for this session.
-    // Capture the PID under the lock, then drop the lock before the async stop.
-    let pid_to_stop = {
-        let mut agents = state.agents.write().await;
-        agents
-            .remove(&session_id)
-            .and_then(|mut agent| agent.active_pid.take())
-    };
-    if let Some(pid) = pid_to_stop {
-        let _ = claudette::agent::stop_agent(pid).await;
-    }
-
-    let fresh = db
-        .archive_chat_session_ensuring_active(&session_id, &workspace_id)
-        .map_err(|e| e.to_string())?;
-
-    // Rebuild the tray so per-workspace running/attention aggregates reflect
-    // the removed agent (and, if this was the last session, the auto-created
-    // replacement). Without this, the tray can keep showing stale state until
-    // another action triggers a rebuild.
-    crate::tray::rebuild_tray(&app);
-
-    if let Some(fresh) = fresh {
-        let agents = state.agents.read().await;
-        return Ok(Some(hydrate_session(fresh, &agents)));
-    }
-
-    Ok(None)
-}
-
-#[cfg(test)]
-mod mcp_inject_tests {
-    use super::inject_claudette_mcp_entry;
-    use claudette::agent_mcp::bridge::BridgeHandle;
-
-    fn handle() -> BridgeHandle {
-        BridgeHandle {
-            socket_addr: "/tmp/cmcp/abc.sock".into(),
-            token: "secret".into(),
-        }
-    }
-
-    #[test]
-    fn inject_into_empty_config_creates_wrapper() {
-        let merged = inject_claudette_mcp_entry(None, &handle())
-            .unwrap()
-            .unwrap();
-        let v: serde_json::Value = serde_json::from_str(&merged).unwrap();
-        assert!(v["mcpServers"]["claudette"]["command"].is_string());
-        let args = v["mcpServers"]["claudette"]["args"].as_array().unwrap();
-        assert_eq!(args[0], "--agent-mcp");
-        let env = &v["mcpServers"]["claudette"]["env"];
-        assert_eq!(env["CLAUDETTE_MCP_SOCKET"], "/tmp/cmcp/abc.sock");
-        assert_eq!(env["CLAUDETTE_MCP_TOKEN"], "secret");
-    }
-
-    #[test]
-    fn inject_preserves_existing_servers() {
-        let base = serde_json::json!({
-            "mcpServers": {
-                "playwright": {"type": "stdio", "command": "npx", "args": ["pw"]}
-            }
-        })
-        .to_string();
-        let merged = inject_claudette_mcp_entry(Some(base), &handle())
-            .unwrap()
-            .unwrap();
-        let v: serde_json::Value = serde_json::from_str(&merged).unwrap();
-        assert!(v["mcpServers"]["playwright"]["command"].is_string());
-        assert_eq!(
-            v["mcpServers"]["claudette"]["env"]["CLAUDETTE_MCP_TOKEN"],
-            "secret"
-        );
-    }
-
-    #[test]
-    fn inject_overwrites_collision_with_claudette_name() {
-        // If a user happened to define an MCP server called `claudette`, our
-        // injected entry takes precedence — it's not user-configurable and
-        // we control the wire-up.
-        let base = serde_json::json!({
-            "mcpServers": {
-                "claudette": {"type": "stdio", "command": "rogue"}
-            }
-        })
-        .to_string();
-        let merged = inject_claudette_mcp_entry(Some(base), &handle())
-            .unwrap()
-            .unwrap();
-        let v: serde_json::Value = serde_json::from_str(&merged).unwrap();
-        assert_eq!(v["mcpServers"]["claudette"]["args"][0], "--agent-mcp");
-        assert_ne!(v["mcpServers"]["claudette"]["command"], "rogue");
-    }
-
-    #[test]
-    fn inject_rejects_malformed_base_json() {
-        let res = inject_claudette_mcp_entry(Some("not-json".into()), &handle());
-        assert!(res.is_err());
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use super::{RequestedFlags, SessionFlags, persistent_session_flags_drifted};
+    use super::{
+        RequestedFlags, SessionFlags, build_permission_response, persistent_session_flags_drifted,
+    };
 
     fn s(values: &[&str]) -> Vec<String> {
         values.iter().map(|v| (*v).to_string()).collect()
@@ -3085,7 +1822,6 @@ mod tests {
         ));
     }
 
-    use super::build_permission_response;
     use serde_json::json;
 
     #[test]
@@ -3156,84 +1892,5 @@ mod tests {
         let input = json!({});
         let response = build_permission_response(&s(&["*", "Read"]), false, false, "Edit", &input);
         assert_eq!(response["behavior"], "deny");
-    }
-
-    use super::take_stop_snapshot;
-    use crate::state::AgentSessionState;
-    use std::collections::HashMap;
-
-    fn fresh_session(session_id: &str, turn_count: u32, pid: Option<u32>) -> AgentSessionState {
-        AgentSessionState {
-            workspace_id: String::new(),
-            session_id: session_id.to_string(),
-            turn_count,
-            active_pid: pid,
-            custom_instructions: None,
-            needs_attention: false,
-            attention_kind: None,
-            attention_notification_sent: false,
-            persistent_session: None,
-            mcp_config_dirty: false,
-            session_plan_mode: false,
-            session_allowed_tools: Vec::new(),
-            session_disable_1m_context: false,
-            pending_permissions: HashMap::new(),
-            session_exited_plan: false,
-            session_resolved_env: Default::default(),
-            mcp_bridge: None,
-            last_user_msg_id: None,
-            posted_env_trust_warning: false,
-        }
-    }
-
-    #[test]
-    fn take_stop_snapshot_preserves_session_identity_for_resume() {
-        // Regression guard for the bug where Stop wiped session_id/turn_count,
-        // causing the next send to start a brand-new conversation instead of
-        // resuming. Stop must leave those fields intact so send_chat_message
-        // can spawn the CLI with --resume <session_id>.
-        let mut session = fresh_session("sess-abc", 7, Some(12345));
-
-        let (drained, pid, ended_sid) = take_stop_snapshot(&mut session);
-
-        // Caller receives the pid to kill and the sid to log.
-        assert!(drained.is_none(), "no pending permissions to drain");
-        assert_eq!(pid, Some(12345));
-        assert_eq!(ended_sid.as_deref(), Some("sess-abc"));
-
-        // Session identity is preserved — this is the fix.
-        assert_eq!(session.session_id, "sess-abc");
-        assert_eq!(session.turn_count, 7);
-        assert!(session.persistent_session.is_none());
-
-        // active_pid is consumed so the caller can kill without racing.
-        assert!(session.active_pid.is_none());
-    }
-
-    #[test]
-    fn take_stop_snapshot_is_idempotent_when_already_stopped() {
-        // A double-stop (e.g. user clicks Stop twice) must not corrupt state.
-        let mut session = fresh_session("sess-abc", 7, None);
-
-        let (_, pid, ended_sid) = take_stop_snapshot(&mut session);
-
-        assert!(pid.is_none());
-        assert_eq!(ended_sid.as_deref(), Some("sess-abc"));
-        assert_eq!(session.session_id, "sess-abc");
-        assert_eq!(session.turn_count, 7);
-    }
-
-    #[test]
-    fn take_stop_snapshot_clears_attention_flags() {
-        use crate::state::AttentionKind;
-
-        let mut session = fresh_session("sess-abc", 3, Some(999));
-        session.needs_attention = true;
-        session.attention_kind = Some(AttentionKind::Ask);
-
-        take_stop_snapshot(&mut session);
-
-        assert!(!session.needs_attention);
-        assert!(session.attention_kind.is_none());
     }
 }
