+++ src-tauri/src/main.rs
@@ -473,26 +473,26 @@ fn main() {
             commands::files::open_attachment_in_browser,
             commands::files::open_attachment_with_default_app,
             // Chat
-            commands::chat::load_chat_history,
-            commands::chat::load_attachments_for_session,
-            commands::chat::load_attachment_data,
-            commands::chat::read_file_as_base64,
-            commands::chat::send_chat_message,
-            commands::chat::stop_agent,
-            commands::chat::reset_agent_session,
-            commands::chat::clear_attention,
-            commands::chat::submit_agent_answer,
-            commands::chat::submit_plan_approval,
-            commands::chat::list_checkpoints,
-            commands::chat::rollback_to_checkpoint,
-            commands::chat::clear_conversation,
-            commands::chat::save_turn_tool_activities,
-            commands::chat::load_completed_turns,
-            commands::chat::list_chat_sessions,
-            commands::chat::get_chat_session,
-            commands::chat::create_chat_session,
-            commands::chat::rename_chat_session,
-            commands::chat::archive_chat_session,
+            commands::chat::send::load_chat_history,
+            commands::chat::send::send_chat_message,
+            commands::chat::attachments::load_attachments_for_session,
+            commands::chat::attachments::load_attachment_data,
+            commands::chat::attachments::read_file_as_base64,
+            commands::chat::lifecycle::stop_agent,
+            commands::chat::lifecycle::reset_agent_session,
+            commands::chat::interaction::clear_attention,
+            commands::chat::interaction::submit_agent_answer,
+            commands::chat::interaction::submit_plan_approval,
+            commands::chat::checkpoint::list_checkpoints,
+            commands::chat::checkpoint::rollback_to_checkpoint,
+            commands::chat::checkpoint::clear_conversation,
+            commands::chat::checkpoint::save_turn_tool_activities,
+            commands::chat::checkpoint::load_completed_turns,
+            commands::chat::session::list_chat_sessions,
+            commands::chat::session::get_chat_session,
+            commands::chat::session::create_chat_session,
+            commands::chat::session::rename_chat_session,
+            commands::chat::session::archive_chat_session,
             // Plan
             commands::plan::read_plan_file,
             // Metrics
